pyproject.toml
[project]
name = "pai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    # TensorFlow: Necessário para o código de deep learning (Inception)
    "tensorflow>=2.15.0,<2.16.0",
    # NumPy: Versão compatível com TensorFlow (2.x ainda não suportado)
    "numpy>=1.26.0,<1.27.0",
    # Keras removido - usaremos tf.keras que vem com TensorFlow
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=10.0.0,<11.0.0",
    "pyqt5>=5.15.11",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "xgboost>=3.0.2",
]
